Exercise 01 oop

+++ src/main/java/oop/ex01/Triangle.java
@@ -6,7 +6,7 @@ public class Triangle {
     private double b;
     private double c;
 
-    public Triangle(int x1, int y1, int x2, int y2, int x3, int y3) {
+    public Triangle(double x1, double y1, double x2, double y2, double x3, double y3) {
         this.a = getDistance(x1, y1, x2, y2);
         this.b = getDistance(x1, y1, x3, y3);
         this.c = getDistance(x2, y2, x3, y3);
